Fixed intergration of Models and handlers acc to latest schema
