+++ packages/server/models/org.go
@@ -1,16 +1,11 @@
 package models
 
 type Organization struct {
-	Id           string      `json:"id"`
-	Name         string      `json:"name"`
-	APITokenHash string      `json:"api_token_hash"`
-	JoinCode     string      `json:"join_code"`
-	Users        interface{} `json:"users"`
-}
-
-type OrgAccess struct {
-	Id           string `json:"id"`
-	Name         string `json:"name"`
+	BaseModel
+	Name         string `json:"name" gorm:"not null"`
+	Handle       string `json:"handle" gorm:"unique"`
+	Avatar       string `json:"avatar"`
+	Description  string `json:"description"`
 	APITokenHash string `json:"api_token_hash"`
-	Role         string `json:"role"`
+	JoinCode     string `json:"join_code" gorm:"not null"`
 }
