+++ packages/server/models/response.go
@@ -1,14 +1,27 @@
 package models
 
+import "net/http"
+
+type ResponseBody struct {
+	Status  int
+	Data    interface{}
+	Message string
+}
+
 type Response struct {
 	Error      error
-	Message    string
-	Body       interface{}
+	Body       ResponseBody
 	StatusCode int
 }
 
 func NewErrorResponse(err error) *Response {
 	return &Response{
 		Error: err,
+		Body: ResponseBody{
+			Status:  http.StatusInternalServerError,
+			Message: err.Error(),
+			Data:    nil,
+		},
+		StatusCode: http.StatusInternalServerError,
 	}
 }
