+++ packages/server/service/create_org.go
@@ -10,13 +10,15 @@ import (
 func CreateOrg(request *models.Request) *models.Response {
 	request.ParseJsonBody()
 	orgName := request.GetParsedBodyAttribute("name").(string)
-	mailId := request.User.MailId
-	org, err := datastore.CreateOrgFromMailId(mailId, orgName)
+	email := request.User.Email
+	org, err := datastore.CreateOrgFromEmail(email, orgName)
 	if err != nil {
 		return models.NewErrorResponse(err)
 	}
 	response := &models.Response{}
 	response.StatusCode = http.StatusOK
-	response.Body = []models.Organization{*org}
+	response.Body.Status = response.StatusCode
+	response.Body.Message = "Organization created"
+	response.Body.Data = []models.Organization{*org}
 	return response
 }
